+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/search/RangeSearchAT.java
@@ -382,4 +382,24 @@ public class RangeSearchAT extends AbstractSearchAT {
     public void testRangeMapField4() {
         filter(range("map_1$k1").lower("a1").upper("k1")).check(0);
     }
+
+    @Test
+    public void testRangeMapFieldWithAlias1() {
+        filter(range("string_map$k1").lower("a").upper("z")).check(2);
+    }
+
+    @Test
+    public void testRangeMapFieldWithAlias2() {
+        filter(range("string_map$k1").lower("a1").upper("z9")).check(2);
+    }
+
+    @Test
+    public void testRangeMapFieldWithAlias3() {
+        filter(range("string_map$k1").lower("a1").upper("k9")).check(0);
+    }
+
+    @Test
+    public void testRangeMapFieldWithAlias4() {
+        filter(range("string_map$k1").lower("a1").upper("k1")).check(0);
+    }
 }