+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 3.0.8.1 (Upcoming)
+
+Merged from 2.2.7.0:
+* Fix mapper referenced by alias in sortFields
 
 ## 3.0.8.0 (July 07, 2016)